backend/store-service/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
# Flask-Migrate>=4.0
Flask-Cors>=3.0
PyJWT>=2.0
python-dotenv>=0.21.0
pymysql>=1.0